Whatever

+++ effects.h
@@ -12,7 +12,7 @@ void invert(int *r, int *g, int *b);
 void monochrome(int *r, int *g, int *b, bool do_threshold, int threshold);
 void quantize(int *r, int *g, int *b, int bit_depth);
 void dither(Uint8 *framebuffer, size_t framebuffer_size, int width, int height, int thresh, float brightness);
-void shift(int *r, int *g, int *b, int color_shift);
+void shift(int *r, int *g, int *b, float color_shift);
 void exposure(int *r, int *g, int *b, float exposure_val);
 
 #endif
